Generate the home page and add styling
